[Core] FIXED_POINT define support added

+++ Addons/Prototype/Attack/Runtime/Views/AnimatorAttackViewModule.cs
@@ -1,7 +1,6 @@
 namespace ME.BECS.Attack {
 
     using Views;
-    using Unity.Mathematics;
     
     public class AnimatorAttackViewModule : IViewApplyState {
 
@@ -14,7 +13,7 @@ namespace ME.BECS.Attack {
             var sensor = ent.GetAspect<ME.BECS.Units.UnitAspect>().readComponentRuntime.attackSensor;
             if (sensor.IsAlive() == false) return;
             var attack = sensor.Read<AttackRuntimeFireComponent>();
-            this.animator.SetFloat(attackHash, attack.fireTimer);
+            this.animator.SetFloat(attackHash, (float)attack.fireTimer);
             
         }
